Moving mux to frontend_connectors
