+++ quesma/quesma/dual_write_proxy.go
@@ -16,6 +16,7 @@ import (
 	"quesma/end_user_errors"
 	"quesma/feature"
 	"quesma/frontend_connectors"
+	"quesma/frontend_connectors/mux"
 	"quesma/frontend_connectors/routes"
 	"quesma/ingest"
 	"quesma/logger"
@@ -23,7 +24,6 @@ import (
 	"quesma/quesma/async_search_storage"
 	"quesma/quesma/config"
 	"quesma/quesma/gzip"
-	"quesma/quesma/mux"
 	"quesma/quesma/recovery"
 	"quesma/quesma/types"
 	"quesma/quesma/ui"
