+++ quesma/quesma/router.go
@@ -11,6 +11,7 @@ import (
 	"quesma/elasticsearch"
 	"quesma/end_user_errors"
 	"quesma/frontend_connectors"
+	"quesma/frontend_connectors/mux"
 	"quesma/frontend_connectors/routes"
 	"quesma/ingest"
 	"quesma/logger"
@@ -22,7 +23,6 @@ import (
 	"quesma/quesma/functionality/field_capabilities"
 	"quesma/quesma/functionality/resolve"
 	"quesma/quesma/functionality/terms_enum"
-	"quesma/quesma/mux"
 	"quesma/quesma/types"
 	"quesma/quesma/ui"
 	"quesma/schema"
